requirements.txt
django==3.2.4
Markdown==3.1.1
django-filter==2.0.0
djangorestframework==3.12.2
psycopg2==2.9.9
django-ckeditor==5.9.0
Pillow==7.1.2
itsdangerous==2.1.0